% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{ATEobj}
\alias{ATEobj}
\title{ATEobj}
\usage{
ATEobj(processed_data, constraints)
}
\arguments{
\item{processed_data}{A bcDAT object which is the output of the process_data() function.}

\item{constraints}{A list of constraints which is the output of the make_constraints() function. Only the hyperplane arrangement is used here.}
}
\value{
A numerical vector containing values of the objective function used to bound the ATE.
}
\description{
Constructs the objective function to bound the ATE
}
\examples{
# process_data1 is output from process_data()
# cons0 is output from make_constraints()
objlam_ATE = ATEobj(processed_data = process_data1,constraints = cons0)
}
