% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{gen_perturb}
\alias{gen_perturb}
\title{gen_perturb}
\usage{
gen_perturb(Acons, perturb_support = c(-1e-06, 1e-06, 0, 1e-06))
}
\arguments{
\item{Acons}{defines the linear constraints in the form of Acons%*%x - bcons <= b_n}

\item{obj}{defines the linear coefficients of the linear objective function}
}
\value{
per_b1 and per_b2 is the perturbation of two inequality (implied from the equality constraint) constraints, each of length equals to the number of constraints (i.e. nrows(Acons)). per_psi is the perturbation of the objective, of length equals to the number of variables (i.e. ncol(Acons)). per_r1 and per_r2 is the perturbation on the support of the variables, each of length the number of variables.
}
\description{
generate perturbation to apply Cho and Russell (2020)
}
