% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{getbound}
\alias{getbound}
\title{getbound}
\usage{
getbound(obj, Acons, bcons, verb = 1)
}
\arguments{
\item{obj}{defines the linear coefficients of the linear objective function}

\item{Acons}{defines the linear constraints in the form of Acons%*%x = bcons}

\item{bcons}{defines the linear constraints in the form of Acons%*%x = bcons}

\item{verb}{defines how much information of the optimization routine to be printed. Default is set at 1 for no print, change to 5 for full output print}
}
\value{
bound is the lower and upper bound of the value function, phat is the optimized solution for lower and upper bound, status records the status of the two optimization problem
}
\description{
calls Rmosek to solve for two linear programming problem for the identified set
}
\examples{
 ## min/max_x obj'x s.t. Acons \%*\% x = bcons
}
