% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{getbound_per}
\alias{getbound_per}
\title{getbound_per}
\usage{
getbound_per(obj, Acons, bcons, pervar, verb = 1)
}
\arguments{
\item{obj}{defines the linear coefficients of the linear objective function}

\item{Acons}{defines the linear constraints in the form of Acons\%*\%x = bcons}

\item{pervar}{defines the perturbation as in Cho and Russell (2020).}
}
\value{
optimal value of the objective functions for the lower_minus, lower_plus, upper_minus, upper_plus LP problems. Definition see Cho and Russell (2020). phat are the associated optimal solutions and status is the solution status of the optimization problem.
#' @examples
Suppose the original LP is min/max_x c'x s.t. A\%*\%x = b, Cho and Russell (2020) solves four perturbed LP. Details see Cho and Russell (2020).
}
\description{
Use Rmosek to solve four perturbed LP problems as proposed in Cho and Russell (2020) to get bias-corrected estimator for the identified set.
}
