% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{getbound_per_relax}
\alias{getbound_per_relax}
\title{getbound_per_relax}
\usage{
getbound_per_relax(obj, Acons, bcons, Aextra, bextra, pervar, verb = 1)
}
\arguments{
\item{obj}{defines the linear coefficients of the linear objective function}

\item{Acons}{defines the linear constraints in the form of Acons%*%x = bcons}

\item{bcons}{defines the linear constraints in the form of Acons%*%x = bcons}

\item{Aextra}{defines the relaxed linear equality constraints in the form Aextra %*% x <= bextra and Aextra %% x >= -bextra}

\item{bextra}{defines the amount of relaxation of the equality constraints.}

\item{pervar}{defines the perturbation as in Cho and Russell (2020).}
}
\value{
optimal value of the objective functions for the lower_minus, lower_plus, upper_minus, upper_plus LP problems. Definition see Cho and Russell (2020). phat are the associated optimal solutions and status is the solution status of the optimization problem.
#' @examples
}
\description{
In the original LP, we allow relaxation of the IV assumption: Aextra %*% x = 0 to be
Aextra %*% x <= bextra and
Aextra %*% x >= -bextra
while maintaining the equality constraints Acons %*% x = bcons.
And then we solve the four perturbed LP problems as in Cho and Russell (2020) using Rmosek.
}
\examples{
Original LP is defined as
 ## min/max_x obj'x
 s.t. Acons \%*\% x - bcons = 0
      Aextra \%*\% x <= bextra
      Aextra \%*\% x >= -bextra
We then define the four perturbed LP as in Cho and Russell (2020).
}
