% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{make_constraints}
\alias{make_constraints}
\title{Make Constrains}
\usage{
make_constraints(data, f, eliminate)
}
\arguments{
\item{data}{A bcDAT object - the output from process_data()}

\item{f}{The arrangement sign vector.}

\item{eliminate}{A list containing the monotonicity constraints.}
}
\value{
The A matrix and b vector for the system of linear equations Ax=b
}
\description{
This function constructs constraints and bounds. The behaviour of this function is determined by the formula provided by the user
in the process_data() function.
}
\examples{
## The general case where no functional form is imposed.
fsub <- list(SignVector = t(expand.grid(rep(list(c(-1,1)), nrow(XZsupport)))))
eliset <- list( c(1, 5), c(1,7),c(3, 7),c(3,5))
constraints <- make_constraints(data = processed_data, f = fsub, eliminate = eliset)
}
