% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_data}
\alias{process_data}
\title{Process data}
\usage{
process_data(formula, data, support)
}
\arguments{
\item{formula}{A formula object of the form y ~ x1 + ... + xk | z1 + ... + zk | w1 + ... wk, where x, z, and w denote endogenous, exogenous, and instrumental variables, respectively. To exclude exogenous variables, use -1.}

\item{data}{The full data frame of class "data.frame"}

\item{support}{The full support of the right-hand side variables of class "data.frame"}
}
\value{
An object of class "bcDAT" containing the formula, the full data frame, and the support data frame with attached probability vectors.
}
\description{
This function processes a user-supplied data frame based on a provided support data frame and a user-supplied formula object.
The probability vectors generated by this function are used to generate the constraints in the model.
}
\examples{
 ## Example from Gu and Russell (2021)
XZsupport <- doctor_data \%>\% select(ins,married,healthy,emp) \%>\% unique() \%>\% arrange(ins,married,healthy,emp)
processed_data <- process_data(formula = doctor ~ ins | married + healthy | emp,data = doctor_data, support = XZsupport)
}
