% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{getbound}
\alias{getbound}
\title{getbound}
\usage{
getbound(obj, Acons, bcons, verb = 1)
}
\arguments{
\item{Acons}{description}

\item{bcons}{description}

\item{verb}{description}
}
\value{
something returned
}
\description{
description
}
\examples{
 ## Example here
example code
}
