% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{make_constraints}
\alias{make_constraints}
\title{Make Constrains}
\usage{
make_constraints(data, f, eliminate)
}
\arguments{
\item{data}{A bcDAT object - the output from process_data()}

\item{f}{The arrangement sign vector.}

\item{eliminate}{A list containing the monotonicity constraints.}
}
\value{
The A matrix and b vector for the linear system of equations Ax=b
}
\description{
Description here
}
\examples{

}
