% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{make_constraints}
\alias{make_constraints}
\title{Make Constrains}
\usage{
make_constraints(data, f, eliminate)
}
\arguments{
\item{data}{Path to the input file}

\item{f}{Path to the input file}

\item{eliminate}{Path to the input file}
}
\value{
An object
}
\description{
Description here
}
