% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_data}
\alias{process_data}
\title{Process data}
\usage{
process_data(formula, data, support)
}
\arguments{
\item{formula}{Path to the input file}

\item{data}{Path to the input file}

\item{support}{Path to the input file}
}
\value{
An object
}
\description{
Description here
}
